overwrite flag

+++ convert_csv.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 for file in *.csv
 do
-	sed 's/\t/,/g' "$file" > "c_${file}"
+	if [ "$1" == "--overwrite" ]
+	then
+		sed 's/\t/,/g' "$file" > "temp";
+		mv temp "$file";
+	else
+	
+		sed 's/\t/,/g' "$file" > "c_${file}"
+	fi
 done
